backend/requirements.txt
# FastAPI and server
fastapi==0.115.6
uvicorn[standard]==0.34.0
python-multipart==0.0.18

# Database
sqlalchemy[asyncio]==2.0.36
asyncpg==0.30.0
alembic==1.14.0

# Redis
redis==5.2.1

# Celery (task queue)
celery[redis]==5.3.6

# Authentication
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
pydantic[email]==2.10.3
pydantic-settings==2.7.0

# Email sending (for magic links)
aiosmtplib==3.0.2
emails==0.6

# Gmail API
google-api-python-client==2.154.0
google-auth-oauthlib==1.2.1
google-auth-httplib2==0.2.0

# Claude API
anthropic==0.40.0

# Utilities
python-dotenv==1.0.1
httpx==0.28.1
